Pro/Backend/app/requirements.txt
fastapi==0.110.2
h5py==3.11.0
huggingface-hub==0.23.0
keras==3.3.2
nltk==3.8.1
npm==0.1.1
numpy==1.26.4
pillow==10.3.0
pydantic==2.7.1
pyzmq==26.0.2
scikit-learn==1.4.2
tensorflow==2.16.1
tf_keras==2.16.0
transformers==4.40.2
uvicorn==0.29.0